Other arguments for smbl.utils.shell
